pyproject.toml
[tool.poetry]
name = "jota"
version = "0.1.0"
description = ""
authors = ["woshington <woshington@usezapay.com.br>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.1"
django-restframework = "^0.0.1"
djangorestframework-simplejwt = "^5.5.0"
drf-yasg = "^1.21.10"
python-decouple = "^3.8"
psycopg2 = "^2.9.10"
pillow = "^11.2.1"
celery = "^5.5.2"
django-celery-beat = "^2.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
